+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.0.4
+- Added the possibility to rearrange the buttons
+- Added the possibility to display success messages of backend calls
+
 ## 1.0.3
 - Fixed HA 0.116 compatibility issue (#3)